requirements.txt
discord.py==2.3.2
yt-dlp==2023.7.6
pynacl==1.5.0
python-dotenv==1.0.0
